+++ packages/publisher/vitest.unit.config.ts
@@ -27,6 +27,8 @@ export default defineConfig({
       'test/async-lift-chain-proof-dispatch-2270.test.ts',
       'test/async-lift-reconcile-demand.test.ts',
       'test/async-lift-wallet-release.test.ts',
+      'test/async-lift-receipt-hint.test.ts',
+      'test/abort-boundary.test.ts',
       'test/async-lift-runner.test.ts',
       'test/async-lift-runner-wallet-wake.test.ts',
       'test/pre-broadcast-signal-await.test.ts',
